Add tests for PoolList
